Crate yohan_boogaert_1995

I’m a Former emergency response driver, who always had a big interest in programming and technology. I’ve been running Linux for 6 years, I tried many distributions (Ubuntu, Tails, Manjaro,…). I use Arch Linux as my daily driver because Arch gives me the opportunity to really discover what it’s going on under the hood.

I had the chance to get insight knowledge on Rust in 2020 from a friend that became my mentor. I’ve got the opportunity to learn Rust at RustMinded including all the technical and non-technical skills to become a developer.

You can find me on Github and Twitter. You can also send me an email.

Modules

between_2014_and_2019

Multiples jobs with hands on the wheel (Ambulance, Shuttle, Taxi,…).

between_2019_and_2020

Saving peoples from Covid at Centre de Secours Monceau.

current

First job with hands on the keyboard at RustMinded. Discovering the OSS world when learning how to become an Rustacean and a developer.

Macros

cargo_temp

A CLI tool that allows you to create a temporary new Rust project using cargo with already installed dependencies. [Repository]

create_process_w

A crate to create and handle processes on Windows (similar too std::process). [Repository]

intro_to_rust

My vision of the Rust Programming Language as an introduction to Rust. [Repository]

tui_do

Yet Another To-Do App in your Terminal. (Under development) [Repository]

yewprint_playground

A Yew and Yewprint playground using wasm-run. [Repository]

Structs

CargoBreaking

A CLI tool that compares a crate’s public API between two different branches, shows what changed, and suggests the next version according to semver. [Repository]

RecruitRust

The place to be hired as an awesome Rustacean. (Under development) [Repository]

RustBookFr

French translation of the book “The Rust Programming Language”. [Repository]

ThirdIFrontend

Third-I, the only camera that replicates your hearing and your point of view. [Repository] | [Site]

WasmRun

Bundles and tooling for Rust WASM frontend application. [Repository]

XtaskWasm

Customizable commands based on xtask for your WASM project. [Repository]

Yewprint

Port of blueprintjs to Yew. [Repository]

Enums

OssProjectMaintenance

Release, update and maintenance of an OSS project.

ProjectOrganisation

Project organisation using GitHub Project and GitHub settings.

SettingUpCI

Continuous Integration via GitHub Actions.

WebDeployment

Web application deployment using Netlify.

Constants

FRAMEWORKS

Yew.

OS

Arch Linux, Pop!_OS.

PROGRAMMING_LANGUAGES

Rust, Bash.

SPOKEN_LANGUAGES

French (5. Native), English (3. Minimum Professional Proficiency).

TECHNOLOGIES

Git, WASM, CLI, Terminal UI.

Traits

Autodidact

Enthusiast

Flexible

Passionate

TeamPlayer

WantToBuild

WantToLearn