yohan_boogaert_1995
Module exp01_2022_2025

Freelance mission as a Rust Developer at Ystorian working on a cybersecurity platform that gathers information from the OS, BIOS, kernel or network on different systems in parallel to assess security vulnerabilities and compliance level.