Crate yohan_boogaert_1995

I’m a Former emergency response driver, who always had a big interest in programming and technology. I’ve been running Linux for 6 years, I tried many distributions (Ubuntu, Tails, Manjaro,…). I use Arch Linux as my daily driver because Arch gives me the opportunity to really discover what it’s going on under the hood.

I had the chance to get insight knowledge on Rust in 2020 from a friend that became my mentor. I’ve got the opportunity to learn Rust at RustMinded including all the technical and non-technical skills to become a developer.

You can find me on GitHub, Twitter and Mastodon. You can also send me an email.

Modules

exp01_2022_today

Freelance mission as a Rust Developer at Ystorian working on a cybersecurity platform that gathers information from the OS, BIOS, kernel or network on different systems in parallel to assess security vulnerabilities and compliance level.

exp02_2022

Freelance mission as a Rust Developer at Cumul.io rewriting in Rust a query engine for data analytics using faster and more efficient queries and improving the relation of multiple microservices in an AWS cloud environment.

exp03_2020_2022

First job with hands on the keyboard at RustMinded. Discovering the OSS world when learning how to become an Rustacean.

exp04_2014_2020

Multiple jobs with hands on the wheel (emergency response driver, airport shuttle, taxi,…).

Macros

cargo_temp

A CLI tool that allows you to create a temporary new Rust project using cargo with already installed dependencies. [Repository]

create_process_w

A crate to create and handle processes on Windows (similar too std::process). [Repository]

intro_to_rust

My vision of the Rust Programming Language as an introduction to Rust. [Repository]

tui_do

Yet Another To-Do App in your Terminal. (Under development) [Repository]

yewprint_playground

A Yew and Yewprint playground using wasm-run. [Repository]

Structs

CargoBreaking

A CLI tool that compares a crate’s public API between two different branches, shows what changed, and suggests the next version according to semver. [Repository]

CargoGenerate

A developer tool to help you get up and running quickly with a new Rust project by leveraging a pre-existing git repository as a template. [Repository]

Chrono

Date and Time for Rust. [Repository]

RecruitRust

The place to be hired as an awesome Rustacean. (Under development) [Repository]

RustBookFr

French translation of the book “The Rust Programming Language”. [Repository]

ThirdIFrontend

Third-I, the only camera that replicates your hearing and your point of view. [Repository] | [Site]

TopoJson

TopoJSON bindings and utilities for Rust. [Repository]

WasmRun

Bundles and tooling for Rust WASM frontend application. [Repository]

XtaskWasm

Customizable commands based on xtask for your WASM project. [Repository]

XtaskWatch

A customizable helper to watch for changes in your xtask projects. [Repository]

Yewprint

Port of blueprintjs to Yew. [Repository]

Enums

OssProjectMaintenance

Release, update and maintenance of an OSS project.

ProjectOrganisation

Project organisation using GitHub Project and GitHub settings.

SettingUpCI

Continuous Integration via GitHub Actions.

WebDeployment

Web application deployment using Netlify.

Constants

FRAMEWORKS

Yew.

OS

Arch Linux, Pop!_OS.

PROGRAMMING_LANGUAGES

Rust, Bash.

SPOKEN_LANGUAGES

French (5. Native), English (3. Minimum Professional Proficiency).

TECHNOLOGIES

Git, WASM, CLI, Terminal UI, Platform agnosticism.

Traits

Autodidact

Enthusiast

Flexible

Passionate

TeamPlayer

WantToBuild

WantToLearn