Yohan Boogaert - Rust Developer

I started working with Rust in 2020 at RustMinded, where I built a strong foundation in the language and the open-source ecosystem. Since then, I’ve contributed to WebAssembly and Rust tools, worked on rewriting a data query engine to improve performance, or helped on the development of a cybersecurity platform for analyzing vulnerabilities on multiple systems. Along the way, I’ve gained practical experience with JavaScript and TypeScript, working in team environments to deliver effective solutions. I’m also an active contributor to open-source projects and rely on Linux to streamline my development workflow.

Experiences

2022 - 2025

Freelance mission as a Rust Developer at Ystorian working on a cybersecurity platform that gathers information from the OS, BIOS, kernel or network on different systems in parallel to assess security vulnerabilities and compliance level.

2022

Freelance mission as a Rust Developer at Cumul.io rewriting in Rust a query engine for data analytics using faster and more efficient queries and improving the relation of multiple microservices in an AWS cloud environment.

2020 - Current

First job with hands on the keyboard at RustMinded. Discovering the OSS world when learning how to become an Rustacean. Now working as Rust Developer Consultant.

2014 - 2020

Multiple jobs with hands on the wheel (emergency response driver, airport shuttle, taxi,…).

Personal Project

cargo-temp

A CLI tool that allows you to create a temporary new Rust project using cargo with already installed dependencies.

CreateProcessW

A library that provides an API similar to std::process to create and handle processes on Windows using the Win32 API.

yewprint-playground

A Yew and Yewprint playground using wasm-run.

OSS Contributions

cargo-generate

A developer tool to help you get up and running quickly with a new Rust project by leveraging a pre-existing git repository as a template.

chrono

Date and Time for Rust.

gptman

A GPT manager that allows you to copy partitions from one disk to another.

implicit-clone

Immutable types and ImplicitClone trait similar to Copy.

rust-book-fr

French translation of the book “The Rust Programming Language”.

sysinfo

A library to get system information such as processes, Cpus, disks, components and networks.

third-i-frontend

Third-I, the only camera that replicates your hearing and your point of view.

topojson

TopoJSON bindings and utilities for Rust.

xtask-wasm

Customizable commands based on xtask for your WASM project.

xtask-watch

A customizable helper to watch for changes in your xtask projects.

yewprint

A port of blueprintjs to Yew.

Non-technical Skills

Continuous Integration

Continuous Integration via GitHub Actions.

OSS Project Maintenance

Release, update and maintenance of an OSS project.

Project Organisation

Project organisation using GitHub Project and GitHub settings.

Web Deployment

Web application deployment using Netlify.

Technical Skills

Frameworks

Yew, Rocket, Axum, Actix, Tokio, sqlx.

OS

Nixos, Arch Linux, Pop!_OS, macOS, Windows, Debian, Tails.

Programming Languages

Rust, Bash, HTML, CSS, JavaScript.

Spoken Languages

French (5. Native), English (3. Minimum Professional Proficiency), Dutch (1. Basic).

Technologies

Git, WebAssembly, CLI, Terminal UI, Platform agnosticism, System, Web, Backend.

Traits

Autodidact

Enthusiast

Flexible

Passionate

Team player

Want to build

Want to learn