Crate yohan_boogaert_1995

I’m a Former emergency response driver, who always had a big interest in programming and technology. I’ve been running Linux for 6 years, I tried many distributions (Ubuntu, Tails, Manjaro,…). I use Arch Linux as my daily driver because Arch gives me the opportunity to really discover what it’s going on under the hood.

I had the chance to get insight knowledge on Rust in 2020 from a friend that became my mentor. I’ve got the opportunity to learn Rust at RustMinded including all the technical and non-technical skills to become a developer.

You can find me on Github and Twitter. You can also send me an email.

Modules

emergency_medical_technician
Volunteer at Centre de Secours Monceau - 2020

professional_driver
Ambulance, Shuttle, Taxi,… - 2014 to 2019

software_developer
Junior at RustMinded - Current

Structs

CargoTemp
A CLI tool that allows you to create a temporary new Rust project using cargo with already installed dependencies. [Repository]

IntroToRust
My vision of the Rust Programming Language as an introduction to Rust. [Repository]

YewprintPlayground
A Yew and Yewprint playground using wasmbl. [Repository]

Enums

CargoBreaking
A CLI tool that compares a crate’s public API between two different branches, shows what changed, and suggests the next version according to semver. [Repository]

RustBookFr
French translation of the book “The Rust Programming Language”. [Repository]

ThirdiFrontend
Third-I, the only camera that replicates your hearing and your point of view. [Repository] | [Site]

Wasmbl
Bundles and tooling for Rust WASM frontend application. [Repository]

Yewprint
Port of blueprintjs.com to Yew. [Repository]

Constants

FRAMEWORKS
Yew.

OS
Arch Linux, Pop!_OS.

PROGRAMMING_LANGUAGES
Rust, Bash.

SPOKEN_LANGUAGES
French (5. Native), English (3. Minimum Professional Proficiency).

TECHNOLOGIES
Git, WASM, CLI.

Traits

Autodidact

Enthusiast

Flexible

Passionate

TeamPlayer

WantToBuild

WantToLearn